Refactor tweets, photos and posts into concerns
