+++ lib/mingle/concerns.rb
@@ -1,0 +1,3 @@
+module Mingle::Concerns
+  require 'mingle/concerns/models'
+end
