+++ lib/mingle/concerns/models.rb
@@ -1,0 +1,5 @@
+module Mingle::Concerns::Models
+  require 'mingle/concerns/models/facebook'
+  require 'mingle/concerns/models/instagram'
+  require 'mingle/concerns/models/twitter'
+end
