+++ lib/mingle/concerns/models/facebook.rb
@@ -1,0 +1,3 @@
+module Mingle::Concerns::Models::Facebook
+  require 'mingle/concerns/models/facebook/post'
+end
