+++ lib/mingle/concerns/models/instagram.rb
@@ -1,0 +1,3 @@
+module Mingle::Concerns::Models::Instagram
+  require 'mingle/concerns/models/instagram/photo'
+end
