+++ lib/mingle/concerns/models/twitter.rb
@@ -1,0 +1,3 @@
+module Mingle::Concerns::Models::Twitter
+  require 'mingle/concerns/models/twitter/tweet'
+end
